I believe this commit corrects an orbit interpolation error. (#52)
It appears that the variables ii and t2 are used in an iteration to
find the index (ii) of the first orbit velocity entry that occurs at a
time (t2) after the burst time (t_brst). At the end of this iteration
t1 is the time of the last orbit velocity state prior to t_brst, and
the corresponding index is ii-1.

The code following this iteration interpolates vx, vy and vz for
t_brst. However, prior to the change in this commit, if one sets
t_brst to be t1, one would get vx, vy and vz associated with t2, and
vice versa.

I believe this is an error.

Co-authored-by: Matt Calef <mattcalef@descarteslabs.com>