Changes on the fetchOrbit.py to download orbit files from the new ESA website (#123)
I've noticed that this code is similar to fetchOrbit.py in ISCE2 which downloads Sentinel-1 orbit files from ESA Copernicus so I've made changes to the fetchOrbit.py to download the files from the https://scihub.copernicus.eu/gnss/#/home. I've used the generic credentials to be able to query and download the Sentinel-1 orbit files.