Update to correct_insar_with_gnss and extract_one_time_series (#400)
* Update input call in extract_one_time_series.csh

In line 64, it called "scene.tab" by exact name instead of the user-input scene.tab. This replaces the name call with the $5 variable call, so the user-specified scene.tab file name is called (i.e. this introduces more flexibility if a user's scene.tab file is called something slightly different).

* Update correct_insar_with_gnss.csh for upsampling 

Instead of using surface to up-sample the filtered correction grid, I replace it with grdsample. This should help avoid systematic fluctuations in the gridded final product. The surface command is left commented out in case a user wants to try it. Additionally, I remove the redundant grdsample between tmp2 and tmp3 grids.

* Update correct_insar_with_gnss.csh